♻️ Rename and use safer implementation
